Replace existing description input from service settings
