+++ editor.planx.uk/src/pages/FlowEditor/components/Settings/ServiceSettings/index.tsx
@@ -1,7 +1,6 @@
 import Container from "@mui/material/Container";
 import React from "react";
 
-import FlowDescription from "./FlowDescription/FlowDescription";
 import FlowStatus from "./FlowStatus";
 import { FooterLinksAndLegalDisclaimer } from "./FooterLinksAndLegalDisclaimer";
 
@@ -9,7 +8,6 @@ const ServiceSettings: React.FC = () => (
   <Container maxWidth="formWrap">
     <FooterLinksAndLegalDisclaimer />
     <FlowStatus />
-    <FlowDescription />
   </Container>
 );
